fix(network): Removed the need of clientIDResponse packet, since the correct conneceted ID is given when first connecting to the server.

+++ pps/src/main/java/pewpew/smash/game/network/packets/PositionPacket.java
@@ -11,5 +11,5 @@ public class PositionPacket extends BasePacket {
     private int id;
     private int x;
     private int y;
-    private short r; // Rotation degree (0-360)
+    private float r; // Rotation degree (0-360)
 }
